+++ scripts/extract/cz/constants.ts
@@ -12,6 +12,7 @@ export const CZ_REGIONS_DIR = 'cz/regions';
 export const CZ_OBCE_FILE = `${CZ_REGIONS_DIR}/obce.geojson`;
 export const CZ_OKRES_FILE = `${CZ_REGIONS_DIR}/okres.geojson`;
 export const CZ_ZSJ_DIL_NAMES_FILE = `${CZ_REGIONS_DIR}/zsj_dil_names.csv`;
+export const CZ_CISZSJD_FILE = 'cz/raw/dojezdka_national/ciszsjd.geojson';
 
 export const CZ_SOURCE_ID_PROPERTY = 'SOURCE_ID';
 export const CZ_NAME_PROPERTY = 'SOURCE_NAME';